+++ ts/packages/typeagent/src/vector/embeddings.ts
@@ -40,7 +40,7 @@ export function similarity(
     type: SimilarityType,
 ): number {
     if (type === SimilarityType.Cosine) {
-        return vector.cosineSimilarity(x, y);
+        return vector.cosineSimilarityLoop(x, y, vector.euclideanLength(y));
     }
     return vector.dotProduct(x, y);
 }
@@ -69,7 +69,7 @@ export function indexOfNearest(
     } else {
         const otherLen = vector.euclideanLength(other);
         for (let i = 0; i < list.length; ++i) {
-            const score: number = cosineSimilarityLoop(
+            const score: number = vector.cosineSimilarityLoop(
                 list[i],
                 other,
                 otherLen,
@@ -109,7 +109,7 @@ export function indexesOfNearest(
     } else {
         const otherLen = vector.euclideanLength(other);
         for (let i = 0; i < list.length; ++i) {
-            const score: number = cosineSimilarityLoop(
+            const score: number = vector.cosineSimilarityLoop(
                 list[i],
                 other,
                 otherLen,
@@ -120,67 +120,6 @@ export function indexesOfNearest(
         }
     }
     return matches.byRank();
-}
-
-/**
- * A faster unrolled version of cosine similarity designed to run in a loop
- * When possible, use NormalizedEmbeddings and Dot products instead.
- * To normalize your embedding, call createNormalized(...)
- * @param x
- * @param other
- * @param otherLen Magnitude of other
- * @returns
- */
-export function cosineSimilarityLoop(
-    x: vector.Vector,
-    other: vector.Vector,
-    otherLen: number,
-): number {
-    if (x.length != other.length) {
-        throw new Error("Array length mismatch");
-    }
-
-    const len = x.length;
-    const unrolledLength = len - (len % 4);
-    let dotSum = 0;
-    let lenXSum = 0;
-    let i = 0;
-    while (i < unrolledLength) {
-        const xVal0 = x[i];
-        const yVal0 = other[i];
-        const xVal1 = x[i + 1];
-        const yVal1 = other[i + 1];
-        const xVal2 = x[i + 2];
-        const yVal2 = other[i + 2];
-        const xVal3 = x[i + 3];
-        const yVal3 = other[i + 3];
-
-        dotSum += xVal0 * yVal0;
-        dotSum += xVal1 * yVal1;
-        dotSum += xVal2 * yVal2;
-        dotSum += xVal3 * yVal3;
-
-        lenXSum += xVal0 * xVal0;
-        lenXSum += xVal1 * xVal1;
-        lenXSum += xVal2 * xVal2;
-        lenXSum += xVal3 * xVal3;
-
-        i += 4;
-    }
-
-    while (i < len) {
-        const xVal = x[i];
-        const yVal = other[i];
-
-        dotSum += xVal * yVal;
-        lenXSum += xVal * xVal;
-
-        ++i;
-    }
-
-    // Cosine Similarity of X, Y
-    // Sum(X * Y) / |X| * |Y|
-    return dotSum / (Math.sqrt(lenXSum) * otherLen);
 }
 
 export interface TopNList<T> {
